update readme links
github sanitizes html for safety purposed and does not allow target or rel to be used to have links open in a new tab. This commit cleans up the links and gets rid of an unneeded discussions sections under Support. Discussion is through telegram and access is gained by request.